Renamed App.vue -> app.vue
