+++ src/components/course_admin/manage_projects/single_project.vue
@@ -91,7 +91,7 @@ import { Component, Inject, Prop, Vue } from 'vue-property-decorator';
 
 import { Course, Project, User } from 'ag-client-typescript';
 
-import { GlobalData } from '@/App.vue';
+import { GlobalData } from '@/app.vue';
 import APIErrors from "@/components/api_errors.vue";
 import Dropdown from '@/components/dropdown.vue';
 import Modal from '@/components/modal.vue';
