+++ src/styles/global.scss
@@ -1,5 +1,5 @@
 // This file is for SCSS variables, mixins, etc. that need to be available globally.
-// It is NOT for global style rules. Those should go in App.vue
+// It is NOT for global style rules. Those should go in app.vue
 
 @import '@/styles/colors.scss';
